Move frameworks tests (#7318)
