+++ src/frameworks/compose/discover/filesystem.spec.ts
